requirements.txt
google-api-python-client
matplotlib
coloredlogs
python-dateutil
discord.py
loguru
traceback_with_variables
pytz
pytchat
pillow
